module restructure
