+++ crates/sozo/src/commands/migrate.rs
@@ -1,20 +1,16 @@
 use std::env::{self, current_dir};
 
-use anyhow::{anyhow, Context, Result};
+use anyhow::Result;
 use camino::Utf8PathBuf;
 use clap::Args;
 use dotenv::dotenv;
 use scarb::core::Config;
 use scarb::ops;
 use scarb::ui::Verbosity;
-use starknet::accounts::{Account, ConnectedAccount};
-use starknet::core::types::{BlockId, BlockTag, StarknetError};
-use starknet::providers::{Provider, ProviderError};
 
-use super::build::{self, BuildArgs, ProfileSpec};
+use crate::commands::build::{self, BuildArgs, ProfileSpec};
+use crate::ops::migration;
 use crate::ops::migration::config::{EnvironmentConfig, WorldConfig};
-use crate::ops::migration::strategy::{execute_migration, prepare_for_migration};
-use crate::ops::migration::world::WorldDiff;
 
 #[derive(Args)]
 pub struct MigrateArgs {
@@ -29,8 +25,6 @@ pub struct MigrateArgs {
     profile_spec: ProfileSpec,
 }
 
-// TODO: add verbose flag
-// TODO: output the migration plan before executing it
 pub fn run(args: MigrateArgs) -> Result<()> {
     dotenv().ok();
 
@@ -68,43 +62,11 @@ pub fn run(args: MigrateArgs) -> Result<()> {
     let world_config = WorldConfig::from_workspace(&ws).unwrap_or_default();
     let env_config = EnvironmentConfig::from_workspace(profile.as_str(), &ws)?;
 
-    let migration_output = ws.config().tokio_handle().block_on(async {
-        let migrator =
-            env_config.migrator().await.with_context(|| "Failed to initialize migrator account")?;
-
-        migrator
-            .provider()
-            .get_class_hash_at(BlockId::Tag(BlockTag::Pending), migrator.address())
-            .await
-            .map_err(|e| match e {
-                ProviderError::StarknetError(StarknetError::ContractNotFound) => {
-                    anyhow!("Migrator account doesn't exist: {:#x}", migrator.address())
-                }
-                _ => anyhow!(e),
-            })?;
-
-        config.ui().print("🔍 Building world state...");
-
-        let diff = WorldDiff::from_path(target_dir.clone(), &world_config, &env_config).await?;
-        let mut migration = prepare_for_migration(target_dir, diff, world_config)?;
-
-        config.ui().print("🌎 Migrating world...");
-
-        execute_migration(&mut migration, migrator)
-            .await
-            .map_err(|e| anyhow!(e))
-            .with_context(|| "Failed to migrate")
-    })?;
-
-    config.ui().print(format!(
-        "\n✨ Successfully migrated world at address {:#x}",
-        migration_output
-            .world
-            .as_ref()
-            .map(|o| o.contract_address)
-            .or(world_config.address)
-            .expect("world address must exist"),
-    ));
+    ws.config().tokio_handle().block_on(migration::execute(
+        world_config,
+        env_config,
+        &ws.config(),
+    ))?;
 
     Ok(())
 }
