+++ crates/sozo/src/ops/migration/migration_test.rs
@@ -1,8 +1,11 @@
 use camino::Utf8PathBuf;
 use dojo_test_utils::sequencer::Sequencer;
+use scarb::core::Config;
+use scarb::ui::Verbosity;
 
 use crate::ops::migration::config::{EnvironmentConfig, WorldConfig};
-use crate::ops::migration::strategy::{execute_migration, prepare_for_migration};
+use crate::ops::migration::execute_strategy;
+use crate::ops::migration::strategy::prepare_for_migration;
 use crate::ops::migration::world::WorldDiff;
 
 #[tokio::test]
@@ -22,8 +25,13 @@ async fn test_migration() {
         .await
         .unwrap();
 
+    let config = Config::builder(Utf8PathBuf::from_path_buf("../../examples/ecs/".into()).unwrap())
+        .ui_verbosity(Verbosity::Normal)
+        .build()
+        .unwrap();
+
     let mut migration = prepare_for_migration(target_dir, world, WorldConfig::default()).unwrap();
-    execute_migration(&mut migration, env_config.migrator().await.unwrap()).await.unwrap();
+    execute_strategy(&mut migration, env_config.migrator().await.unwrap(), &config).await.unwrap();
 
     sequencer.stop().unwrap();
 }
