+++ crates/sozo/src/ops/migration/strategy.rs
@@ -1,27 +1,17 @@
 use std::collections::HashMap;
 use std::fs;
 use std::path::PathBuf;
-use std::time::Duration;
 
 use anyhow::{anyhow, Context, Result};
 use camino::Utf8PathBuf;
-use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
-use starknet::accounts::ConnectedAccount;
-use starknet::core::types::{FieldElement, InvokeTransactionResult};
-use starknet::providers::Provider;
+use starknet::core::types::FieldElement;
 
 use super::config::WorldConfig;
-use super::object::{
-    ClassMigration, ContractMigration, Declarable, DeployOutput, Deployable, MigrationError,
-    RegisterOutput, WorldContract,
-};
+use super::object::{ClassMigration, ContractMigration, DeployOutput};
+use super::object::{MigrationError, RegisterOutput};
 use super::world::{ClassDiff, ContractDiff, WorldDiff};
 
 pub type MigrationResult<S, P> = Result<MigrationOutput, MigrationError<S, P>>;
-
-#[cfg(test)]
-#[path = "strategy_test.rs"]
-mod test;
 
 #[derive(Debug)]
 pub struct MigrationOutput {
@@ -41,151 +31,12 @@ pub struct MigrationStrategy {
 }
 
 impl MigrationStrategy {
-    fn world_address(&self) -> Option<FieldElement> {
+    pub fn world_address(&self) -> Option<FieldElement> {
         match &self.world {
             Some(c) => c.contract_address,
             None => self.world_config.address,
         }
     }
-}
-
-pub async fn execute_migration<A>(
-    strategy: &mut MigrationStrategy,
-    migrator: A,
-) -> MigrationResult<A::SignError, <A::Provider as Provider>::Error>
-where
-    A: ConnectedAccount + Sync,
-{
-    let multi_progress = MultiProgress::new();
-
-    let executor_output = match &mut strategy.executor {
-        Some(executor) => {
-            let eb = ProgressBar::new_spinner()
-                .with_style(ProgressStyle::with_template("{spinner} executor: {msg}").unwrap());
-            eb.enable_steady_tick(Duration::from_millis(100));
-
-            multi_progress.add(eb.clone());
-
-            eb.set_message("deploying contract");
-            let res = executor.deploy(vec![], &migrator).await?;
-            eb.finish_with_message("deployed");
-
-            if strategy.world.is_none() {
-                let addr = strategy.world_address().ok_or(MigrationError::WorldAddressNotFound)?;
-                WorldContract::new(addr, &migrator).set_executor(res.contract_address).await?;
-            }
-
-            Some(res)
-        }
-        None => None,
-    };
-
-    let world_output = match &mut strategy.world {
-        Some(world) => {
-            let wb = ProgressBar::new_spinner()
-                .with_style(ProgressStyle::with_template("{spinner} world: {msg}").unwrap());
-            wb.enable_steady_tick(Duration::from_millis(100));
-            multi_progress.add(wb.clone());
-
-            wb.set_message("deploying contract");
-
-            let res = world
-                .deploy(
-                    vec![strategy.executor.as_ref().unwrap().contract_address.unwrap()],
-                    &migrator,
-                )
-                .await?;
-
-            wb.finish_with_message("deployed");
-
-            Some(res)
-        }
-        None => None,
-    };
-
-    let components_output = register_components(&strategy, &migrator, &multi_progress).await?;
-    let systems_output = register_systems(&strategy, &migrator, &multi_progress).await?;
-
-    multi_progress.clear().expect("should be able to clear progress bar");
-
-    Ok(MigrationOutput {
-        world: world_output,
-        executor: executor_output,
-        systems: systems_output,
-        components: components_output,
-    })
-}
-
-async fn register_components<A>(
-    strategy: &MigrationStrategy,
-    migrator: &A,
-    multi_progress: &MultiProgress,
-) -> Result<RegisterOutput, MigrationError<A::SignError, <A::Provider as Provider>::Error>>
-where
-    A: ConnectedAccount + Sync,
-{
-    let cb = ProgressBar::new_spinner()
-        .with_style(ProgressStyle::with_template("{spinner} components: {msg}").unwrap());
-    cb.enable_steady_tick(Duration::from_millis(100));
-
-    multi_progress.add(cb.clone());
-
-    let total = strategy.components.len();
-    let mut declare_output = vec![];
-
-    for (i, component) in strategy.components.iter().enumerate() {
-        cb.set_message(format!("({i}/{total}) declaring {} class", component.class.name));
-        let res = component.declare(migrator).await?;
-        declare_output.push(res);
-    }
-
-    let world_address = strategy.world_address().ok_or(MigrationError::WorldAddressNotFound)?;
-
-    cb.set_message(format!("registering components to world"));
-    let InvokeTransactionResult { transaction_hash } = WorldContract::new(world_address, migrator)
-        .register_components(&declare_output.iter().map(|o| o.class_hash).collect::<Vec<_>>())
-        .await?;
-
-    cb.set_message("registered");
-    cb.finish();
-
-    Ok(RegisterOutput { transaction_hash, declare_output })
-}
-
-async fn register_systems<A>(
-    strategy: &MigrationStrategy,
-    migrator: &A,
-    multi_progress: &MultiProgress,
-) -> Result<RegisterOutput, MigrationError<A::SignError, <A::Provider as Provider>::Error>>
-where
-    A: ConnectedAccount + Sync,
-{
-    let sb = ProgressBar::new_spinner()
-        .with_style(ProgressStyle::with_template("{spinner} systems: {msg}").unwrap());
-    sb.enable_steady_tick(Duration::from_millis(100));
-
-    multi_progress.add(sb.clone());
-
-    let total = strategy.systems.len();
-    let mut declare_output = vec![];
-
-    for (i, system) in strategy.systems.iter().enumerate() {
-        sb.set_message(format!("({i}/{total}) declaring {} class", system.class.name));
-        let res = system.declare(migrator).await?;
-        declare_output.push(res);
-    }
-
-    let world_address = strategy.world_address().ok_or(MigrationError::WorldAddressNotFound)?;
-
-    sb.set_message("registering systems to world...");
-
-    let InvokeTransactionResult { transaction_hash } = WorldContract::new(world_address, migrator)
-        .register_systems(&declare_output.iter().map(|o| o.class_hash).collect::<Vec<_>>())
-        .await?;
-
-    sb.finish_with_message("registered");
-
-    Ok(RegisterOutput { transaction_hash, declare_output })
 }
 
 /// construct migration strategy
